- ShellProperties now correctly extract name annotations in () - Improved color and flags handling - Improved mutex handling - Bugfix for thumbnail transparency - Bugfix for LFTypeColor properties - Added new file extensions - Minor edits

+++ LFCommDlg/CInspectorGrid.h
@@ -155,7 +155,7 @@ public:
 	virtual BOOL OnPushChar(UINT nChar);
 
 protected:
-	void OnSetColor(UINT Color);
+	void OnSetColor(BYTE Color);
 
 	static CIcons m_ColorDots;
 };
